+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Unreleased` header.
 - On Orbital, implement `is_decorated`.
 - On Orbital, implement `set_window_level`.
 - On Orbital, emit `DeviceEvent::MouseMotion`.
+- On Wayland, fix title in CSD not updated from `AboutToWait`.
 
 # 0.29.10